+++ Changelog.md
@@ -1,6 +1,7 @@
 ## Changelog
 v 2.78
-  Added Cup image for Safe bonus game
+  Added Cup image for Safe bonus game,
+  BannerCarousel adjustmenets
 
 v 2.77
   Added BannerCarousel shuffle and label description on hover